chore: work on types based on dataset contract
